pyproject.toml
[project]
name = "strands-multi-agent-blog-example"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "strands-agents>=1.13.0",
    "strands-agents-tools>=0.2.12",
]
